Address review on the call recorder transcript provider
- Move RecallAsyncTranscriptProvider into its own type file.
- Use the in operator instead of Object.keys().includes for provider validation.
- Move the manifest SELECT option drift guard out of the util test into an
  application-config test.
