+++ packages/twenty-apps/public/call-recorder/src/logic-functions/constants/default-call-recorder-transcript-provider.ts
@@ -1,4 +1,4 @@
-import { type RecallAsyncTranscriptProvider } from 'src/logic-functions/constants/recall-async-transcript-providers';
+import { type RecallAsyncTranscriptProvider } from 'src/logic-functions/types/recall-async-transcript-provider.type';
 
 export const DEFAULT_CALL_RECORDER_TRANSCRIPT_PROVIDER: RecallAsyncTranscriptProvider =
   'recallai_async';
