+++ packages/twenty-apps/public/call-recorder/src/logic-functions/constants/recall-async-transcript-providers.ts
@@ -2,6 +2,3 @@ export const RECALL_ASYNC_TRANSCRIPT_PROVIDERS = {
   recallai_async: { language_code: 'auto' },
   gladia_v2_async: { language_config: { code_switching: true } },
 } as const;
-
-export type RecallAsyncTranscriptProvider =
-  keyof typeof RECALL_ASYNC_TRANSCRIPT_PROVIDERS;
