+++ packages/twenty-apps/public/call-recorder/src/logic-functions/types/recall-async-transcript-provider.type.ts
@@ -1,0 +1,4 @@
+import { type RECALL_ASYNC_TRANSCRIPT_PROVIDERS } from 'src/logic-functions/constants/recall-async-transcript-providers';
+
+export type RecallAsyncTranscriptProvider =
+  keyof typeof RECALL_ASYNC_TRANSCRIPT_PROVIDERS;
