+++ packages/twenty-apps/public/call-recorder/src/logic-functions/utils/__tests__/get-recall-async-transcript-provider.test.ts
@@ -1,8 +1,6 @@
 import { afterEach, beforeEach, describe, expect, it } from 'vitest';
 
-import applicationConfig from 'src/application-config';
 import { CALL_RECORDER_TRANSCRIPT_PROVIDER_ENV_VAR_NAME } from 'src/logic-functions/constants/call-recorder-transcript-provider-env-var-name';
-import { RECALL_ASYNC_TRANSCRIPT_PROVIDERS } from 'src/logic-functions/constants/recall-async-transcript-providers';
 import { getRecallAsyncTranscriptProvider } from 'src/logic-functions/utils/get-recall-async-transcript-provider.util';
 
 describe('getRecallAsyncTranscriptProvider', () => {
@@ -27,16 +25,5 @@ describe('getRecallAsyncTranscriptProvider', () => {
     expect(getRecallAsyncTranscriptProvider()).toEqual({
       recallai_async: { language_code: 'auto' },
     });
-  });
-
-  it('offers every supported provider as a select option', () => {
-    const options =
-      applicationConfig.config.applicationVariables?.[
-        CALL_RECORDER_TRANSCRIPT_PROVIDER_ENV_VAR_NAME
-      ]?.options ?? [];
-
-    expect(options.map((option) => option.value).sort()).toEqual(
-      Object.keys(RECALL_ASYNC_TRANSCRIPT_PROVIDERS).sort(),
-    );
   });
 });
