+++ packages/twenty-apps/public/call-recorder/src/logic-functions/utils/get-recall-async-transcript-provider.util.ts
@@ -1,17 +1,14 @@
 import { CALL_RECORDER_TRANSCRIPT_PROVIDER_ENV_VAR_NAME } from 'src/logic-functions/constants/call-recorder-transcript-provider-env-var-name';
 import { DEFAULT_CALL_RECORDER_TRANSCRIPT_PROVIDER } from 'src/logic-functions/constants/default-call-recorder-transcript-provider';
-import {
-  RECALL_ASYNC_TRANSCRIPT_PROVIDERS,
-  type RecallAsyncTranscriptProvider,
-} from 'src/logic-functions/constants/recall-async-transcript-providers';
+import { RECALL_ASYNC_TRANSCRIPT_PROVIDERS } from 'src/logic-functions/constants/recall-async-transcript-providers';
+import { type RecallAsyncTranscriptProvider } from 'src/logic-functions/types/recall-async-transcript-provider.type';
 import { getApplicationVariableValue } from 'src/logic-functions/utils/get-application-variable-value.util';
 import { isNonEmptyString } from 'src/logic-functions/utils/is-non-empty-string.util';
 
 const isRecallAsyncTranscriptProvider = (
   value: unknown,
 ): value is RecallAsyncTranscriptProvider =>
-  isNonEmptyString(value) &&
-  Object.keys(RECALL_ASYNC_TRANSCRIPT_PROVIDERS).includes(value);
+  isNonEmptyString(value) && value in RECALL_ASYNC_TRANSCRIPT_PROVIDERS;
 
 export const getRecallAsyncTranscriptProvider = () => {
   const rawValue = getApplicationVariableValue(
